@@ -22,8 +22,8 @@ tasks.compileJava {
 }
 
 java {
-  sourceCompatibility = JavaVersion.VERSION_11
-  targetCompatibility = JavaVersion.VERSION_11
+  sourceCompatibility = JavaVersion.VERSION_16
+  targetCompatibility = JavaVersion.VERSION_16
   modularity.inferModulePath.set(true)
 }
 
