@@ -9,19 +9,19 @@ import de.muspellheim.activitysampling.backend.Event;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
-import lombok.AllArgsConstructor;
-import lombok.NonNull;
-import lombok.Value;
 
-@Value
-@AllArgsConstructor
-public class ActivityLoggedEvent implements Event {
-  @NonNull String id;
-  @NonNull Instant timestamp;
-  @NonNull Duration period;
-  @NonNull String activity;
-  @NonNull List<String> tags;
+public record ActivityLoggedEvent(
+    String id, Instant timestamp, Duration period, String activity, List<String> tags)
+    implements Event {
+  public ActivityLoggedEvent {
+    Objects.requireNonNull(id, "id");
+    Objects.requireNonNull(timestamp, "timestamp");
+    Objects.requireNonNull(period, "period");
+    Objects.requireNonNull(activity, "activity");
+    Objects.requireNonNull(tags, "tags");
+  }
 
   public ActivityLoggedEvent(String id, Instant timestamp, Duration period, String activity) {
     this(id, timestamp, period, activity, List.of());
