@@ -1,8 +1,8 @@
 module de.muspellheim.activitysampling.backend {
-  requires static lombok;
   requires transitive de.muspellheim.activitysampling.contract;
   requires java.logging;
   requires java.prefs;
+  requires static lombok;
   requires org.apache.commons.csv;
 
   exports de.muspellheim.activitysampling.backend;
