@@ -8,13 +8,14 @@ package de.muspellheim.activitysampling.contract.messages.commands;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
-import lombok.NonNull;
-import lombok.Value;
+import java.util.Objects;
 
-@Value
-public class LogActivityCommand {
-  @NonNull LocalDateTime timestamp;
-  @NonNull Duration period;
-  @NonNull String activity;
-  @NonNull List<String> tags;
+public record LogActivityCommand(
+    LocalDateTime timestamp, Duration period, String activity, List<String> tags) {
+  public LogActivityCommand {
+    Objects.requireNonNull(timestamp, "timestamp");
+    Objects.requireNonNull(period, "period");
+    Objects.requireNonNull(activity, "activity");
+    Objects.requireNonNull(tags, "tags");
+  }
 }
