@@ -7,11 +7,11 @@ package de.muspellheim.activitysampling.contract.messages.queries;
 
 import de.muspellheim.activitysampling.contract.data.Activity;
 import java.util.List;
-import lombok.NonNull;
-import lombok.Value;
+import java.util.Objects;
 
-@Value
-public class ActivityLogQueryResult {
-  @NonNull List<Activity> log;
-  @NonNull List<Activity> recent;
+public record ActivityLogQueryResult(List<Activity> log, List<Activity> recent) {
+  public ActivityLogQueryResult {
+    Objects.requireNonNull(log, "log");
+    Objects.requireNonNull(recent, "recent");
+  }
 }
